+++ src/index.js
@@ -1,13 +1,13 @@
 import * as babel from '@babel/core';
 import { createFilter } from 'rollup-pluginutils';
-import preflightCheck from './preflightCheck.js';
-import bundledHelpersPlugin from './bundledHelpersPlugin.js';
-import { addBabelPlugin, escapeRegExpCharacters } from './utils.js';
 import { RUNTIME, EXTERNAL, BUNDLED, INLINE, HELPERS } from './constants.js';
+import bundledHelpersPlugin from './bundledHelpersPlugin.js';
+import preflightCheck from './preflightCheck.js';
+import transformCode from './transformCode';
+import { addBabelPlugin, escapeRegExpCharacters, warnOnce } from './utils.js';
 
 const unpackOptions = ({
 	extensions = babel.DEFAULT_EXTENSIONS,
-	skipBabelHelpersCheck = false,
 	// rollup uses sourcemap, babel uses sourceMaps
 	// just normalize them here so people don't have to worry about it
 	sourcemap = true,
@@ -15,7 +15,18 @@ const unpackOptions = ({
 	sourceMap = true,
 	sourceMaps = true,
 	...rest
-} = {}) => {
+} = {}) => ({
+	extensions,
+	plugins: [],
+	sourceMaps: sourcemap && sourcemaps && sourceMap && sourceMaps,
+	...rest,
+	caller: {
+		name: 'rollup-plugin-babel',
+		...rest.caller,
+	},
+});
+
+const unpackInputPluginOptions = ({ skipBabelHelpersCheck = false, ...rest }) => {
 	if (!rest.babelHelpers) {
 		throw new Error(
 			'You have to specify how do you want to bundle/import "Babel helpers" (runtime functions inserted by Babel which are used by some transformations).\n\n' +
@@ -26,38 +37,61 @@ const unpackOptions = ({
 				`  - "${INLINE}" - this is not recommended. Helpers will be inserted in each file using them, this can cause serious code duplication (this is default Babel behaviour)\n`,
 		);
 	}
-	return {
-		extensions,
-		plugins: [],
-		skipBabelHelpersCheck,
-		sourceMaps: sourcemap && sourcemaps && sourceMap && sourceMaps,
+	return unpackOptions({
 		...rest,
-		caller: { name: 'rollup-plugin-babel', supportsStaticESM: true, supportsDynamicImport: true, ...rest.caller },
-	};
+		skipBabelHelpersCheck,
+		caller: {
+			supportsStaticESM: true,
+			supportsDynamicImport: true,
+			supportsTopLevelAwait: true,
+			...rest.caller,
+		},
+	});
 };
+
+const unpackOutputPluginOptions = (options, { format }) =>
+	unpackOptions({
+		configFile: false,
+		sourceType: format === 'es' ? 'module' : 'script',
+		...options,
+		caller: {
+			supportsStaticESM: format === 'es',
+			...options.caller,
+		},
+	});
+
+function getOptionsWithOverrides(pluginOptions = {}, overrides = {}) {
+	if (!overrides.options) return { customOptions: null, pluginOptionsWithOverrides: pluginOptions };
+	const overridden = overrides.options(pluginOptions);
+
+	if (typeof overridden.then === 'function') {
+		throw new Error(
+			".options hook can't be asynchronous. It should return `{ customOptions, pluginsOptions }` synchronously.",
+		);
+	}
+
+	return {
+		customOptions: overridden.customOptions || null,
+		pluginOptionsWithOverrides: overridden.pluginOptions || pluginOptions,
+	};
+}
 
 const returnObject = () => ({});
 
-function createBabelPluginFactory(customCallback = returnObject) {
+function createBabelInputPluginFactory(customCallback = returnObject) {
 	const overrides = customCallback(babel);
 
 	return pluginOptions => {
-		let customOptions = null;
+		const { customOptions, pluginOptionsWithOverrides } = getOptionsWithOverrides(pluginOptions, overrides);
 
-		if (overrides.options) {
-			const overridden = overrides.options(pluginOptions);
-
-			if (typeof overridden.then === 'function') {
-				throw new Error(
-					".options hook can't be asynchronous. It should return `{ customOptions, pluginsOptions }` synchronously.",
-				);
-			}
-			({ customOptions = null, pluginOptions } = overridden);
-		}
-
-		const { babelHelpers, exclude, extensions, include, skipBabelHelpersCheck, ...babelOptions } = unpackOptions(
-			pluginOptions,
-		);
+		const {
+			exclude,
+			extensions,
+			babelHelpers,
+			include,
+			skipBabelHelpersCheck,
+			...babelOptions
+		} = unpackInputPluginOptions(pluginOptionsWithOverrides);
 
 		const extensionRegExp = new RegExp(`(${extensions.map(escapeRegExpCharacters).join('|')})$`);
 		const includeExcludeFilter = createFilter(include, exclude);
@@ -65,35 +99,20 @@ function createBabelPluginFactory(customCallback = returnObject) {
 
 		return {
 			name: 'babel',
+
 			resolveId(id) {
 				if (id === HELPERS) return id;
 			},
+
 			load(id) {
-				if (id !== HELPERS) {
-					return;
-				}
-
-				return babel.buildExternalHelpers(null, 'module');
+				if (id === HELPERS) return babel.buildExternalHelpers(null, 'module');
 			},
+
 			transform(code, filename) {
-				if (!filter(filename)) return Promise.resolve(null);
-				if (filename === HELPERS) return Promise.resolve(null);
+				if (!filter(filename)) return null;
+				if (filename === HELPERS) return null;
 
-				const config = babel.loadPartialConfig({ ...babelOptions, filename });
-
-				// file is ignored by babel
-				if (!config) {
-					return Promise.resolve(null);
-				}
-
-				return Promise.resolve(
-					!overrides.config
-						? config.options
-						: overrides.config.call(this, config, {
-								code,
-								customOptions,
-						  }),
-				).then(transformOptions => {
+				return transformCode(code, { ...babelOptions, filename }, overrides, customOptions, this, transformOptions => {
 					if (!skipBabelHelpersCheck) {
 						preflightCheck(this, babelHelpers, transformOptions);
 					}
@@ -102,25 +121,74 @@ function createBabelPluginFactory(customCallback = returnObject) {
 						transformOptions = addBabelPlugin(transformOptions, bundledHelpersPlugin);
 					}
 
-					const result = babel.transformSync(code, transformOptions);
-
-					return Promise.resolve(
-						!overrides.result
-							? result
-							: overrides.result.call(this, result, {
-									code,
-									customOptions,
-									config,
-									transformOptions,
-							  }),
-					).then(({ code, map }) => ({ code, map }));
+					return transformOptions;
 				});
 			},
 		};
 	};
 }
 
-const babelPluginFactory = createBabelPluginFactory();
-babelPluginFactory.custom = createBabelPluginFactory;
+function getRecommendedFormat(rollupFormat) {
+	switch (rollupFormat) {
+		case 'amd':
+			return 'amd';
+		case 'iife':
+		case 'umd':
+			return 'umd';
+		case 'system':
+			return 'systemjs';
+		default:
+			return '<module format>';
+	}
+}
+
+function createBabelOutputPluginFactory(customCallback = returnObject) {
+	const overrides = customCallback(babel);
+
+	return pluginOptions => {
+		const { customOptions, pluginOptionsWithOverrides } = getOptionsWithOverrides(pluginOptions, overrides);
+
+		return {
+			name: 'babel',
+
+			renderStart(outputOptions) {
+				const { extensions, include, exclude, allowAllFormats } = pluginOptionsWithOverrides;
+
+				if (extensions || include || exclude) {
+					warnOnce(this, 'The "include", "exclude" and "extensions" options are ignored when transforming the output.');
+				}
+				if (!allowAllFormats && outputOptions.format !== 'es' && outputOptions.format !== 'cjs') {
+					this.error(
+						`Using Babel on the generated chunks is strongly discouraged for formats other than "esm" or "cjs" as it can easily break wrapper code and lead to accidentally created global variables. Instead, you should set "output.format" to "esm" and use Babel to transform to another format, e.g. by adding "presets: [['@babel/env', { modules: '${getRecommendedFormat(
+							outputOptions.format,
+						)}' }]]" to your Babel options. If you still want to proceed, add "allowAllFormats: true" to your plugin options.`,
+					);
+				}
+			},
+
+			renderChunk(code, chunk, outputOptions) {
+				/* eslint-disable no-unused-vars */
+				const {
+					allowAllFormats,
+					exclude,
+					extensions,
+					externalHelpers,
+					externalHelpersWhitelist,
+					include,
+					runtimeHelpers,
+					...babelOptions
+				} = unpackOutputPluginOptions(pluginOptionsWithOverrides, outputOptions);
+				/* eslint-enable no-unused-vars */
+
+				return transformCode(code, babelOptions, overrides, customOptions, this);
+			},
+		};
+	};
+}
+
+const babelPluginFactory = createBabelInputPluginFactory();
+babelPluginFactory.custom = createBabelInputPluginFactory;
+babelPluginFactory.generated = createBabelOutputPluginFactory();
+babelPluginFactory.generated.custom = createBabelOutputPluginFactory;
 
 export default babelPluginFactory;