+++ CHANGELOG.md
@@ -1,4 +1,32 @@
-# rollup-plugin-babel changelog
+## 4.3.2
+
+- Fixed usage with `externalHelpers: true` option
+
+## 4.3.1
+
+- Add `.js` extension to the virtual babel helpers file (only matters when using `preserveModules` option in rollup)
+
+## 4.3.0
+
+- Added `.custom` builder.
+- Fail build when a plugin tries to add non existent babel helper
+
+## 4.2.0
+
+Allow `rollup@1` as peer dependency.
+
+## 4.1.0
+
+- Fixed "preflight check" for ignored files.
+- Return `null` when no transformation has been done (fixing source maps for this case)
+
+## 4.0.3
+
+Fixed fallback class transform in "preflight check".
+
+## 4.0.2
+
+Fixed `rollup` peer dependency.
 
 ## 4.0.0